pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "air-monitor-back"
version = "0.1.0"
description = "Backend for air quality monitoring and forecasting."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.2,<6.0",
  "django-ninja>=1.6,<2.0",
  "PyJWT>=2.10,<3.0",
  "celery>=5.6,<6.0",
  "redis>=5.3,<6.0",
  "psycopg[binary]>=3.3,<4.0",
  "gunicorn>=23.0,<24.0",
  "numpy>=2.2,<3.0",
  "pandas>=2.3,<3.0",
  "torch>=2.8,<3.0",
  "requests>=2.32,<3.0",
  "fake-useragent>=2.2,<3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9.0",
  "pytest-django>=4.11,<5.0",
  "ruff>=0.11,<0.12",
]

[tool.setuptools.packages.find]
include = ["config*", "apps*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
testpaths = ["apps"]
addopts = "-ra --reuse-db"

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
